BE -> grading

+++ backend/internal/adapter/handler/http/features/ujian/list_soal_ujian_for_siswa/list_soal_ujian_siswa_response.go
@@ -5,7 +5,7 @@ type ListSoalUjianSiswaResponse struct {
 	TipeSoal    string                                  `json:"tipe_soal"`
 	Pertanyaan  string                                  `json:"pertanyaan"`
 	Gambar      string                                  `json:"gambar"`
-	BobotSoal   int                                     `json:"bobot_soal"`
+	BobotSoal   float64                                 `json:"bobot_soal"`
 	NoUrutSoal  int                                     `json:"no_urut_soal"`
 	OpsiJawaban []ListSoalUjianSiswaOpsiJawabanResponse `json:"opsi_jawaban"`
 }
